Rename controller to nginx
