+++ hack/verify-flags/exceptions.txt
@@ -1,10 +1,10 @@
-ingress/controllers/nginx-third-party/nginx.tmpl:        require("error_page")
-ingress/controllers/nginx-third-party/nginx.tmpl:    error_page 403 = @custom_403;
-ingress/controllers/nginx-third-party/nginx.tmpl:    error_page 404 = @custom_404;
-ingress/controllers/nginx-third-party/nginx.tmpl:    error_page 405 = @custom_405;
-ingress/controllers/nginx-third-party/nginx.tmpl:    error_page 408 = @custom_408;
-ingress/controllers/nginx-third-party/nginx.tmpl:    error_page 413 = @custom_413;
-ingress/controllers/nginx-third-party/nginx.tmpl:    error_page 501 = @custom_501;
-ingress/controllers/nginx-third-party/nginx.tmpl:    error_page 502 = @custom_502;
-ingress/controllers/nginx-third-party/nginx.tmpl:    error_page 503 = @custom_503;
-ingress/controllers/nginx-third-party/nginx.tmpl:    error_page 504 = @custom_504;
+ingress/controllers/nginx/nginx.tmpl:        require("error_page")
+ingress/controllers/nginx/nginx.tmpl:    error_page 403 = @custom_403;
+ingress/controllers/nginx/nginx.tmpl:    error_page 404 = @custom_404;
+ingress/controllers/nginx/nginx.tmpl:    error_page 405 = @custom_405;
+ingress/controllers/nginx/nginx.tmpl:    error_page 408 = @custom_408;
+ingress/controllers/nginx/nginx.tmpl:    error_page 413 = @custom_413;
+ingress/controllers/nginx/nginx.tmpl:    error_page 501 = @custom_501;
+ingress/controllers/nginx/nginx.tmpl:    error_page 502 = @custom_502;
+ingress/controllers/nginx/nginx.tmpl:    error_page 503 = @custom_503;
+ingress/controllers/nginx/nginx.tmpl:    error_page 504 = @custom_504;
