+++ ingress/controllers/nginx/Makefile
@@ -2,10 +2,10 @@ all: push
 
 # 0.0 shouldn't clobber any release builds
 TAG = 0.4
-PREFIX = gcr.io/google_containers/nginx-third-party
+PREFIX = gcr.io/google_containers/nginx-ingress-controller
 
 controller: controller.go clean
-	CGO_ENABLED=0 GOOS=linux godep go build -a -installsuffix cgo -ldflags '-w' -o nginx-third-party-lb
+	CGO_ENABLED=0 GOOS=linux godep go build -a -installsuffix cgo -ldflags '-w' -o nginx-ingress-controller
 
 container: controller
 	docker build -t $(PREFIX):$(TAG) .
@@ -14,4 +14,4 @@ push: container
 	gcloud docker push $(PREFIX):$(TAG)
 
 clean:
-	rm -f nginx-third-party-lb
+	rm -f nginx-ingress-controller
