+++ ingress/controllers/nginx/examples/rc-custom-configuration.yaml
@@ -1,7 +1,7 @@
 apiVersion: v1
 kind: ReplicationController
 metadata:
-  name: nginx-ingress-3rdpartycfg
+  name: nginx-ingress-controller
   labels:
     k8s-app: nginx-ingress-lb
 spec:
@@ -15,7 +15,7 @@ spec:
         name: nginx-ingress-lb
     spec:
       containers:
-      - image: gcr.io/google_containers/nginx-third-party:0.4
+      - image: gcr.io/google_containers/nginx-ingress-controller:0.4
         name: nginx-ingress-lb
         imagePullPolicy: Always
         livenessProbe:
@@ -49,6 +49,6 @@ spec:
         - containerPort: 8080
           hostPort: 8081
         args:
-        - /nginx-third-party-lb
+        - /nginx-ingress-controller
         - --default-backend-service=default/default-http-backend
         - --nginx-configmap=default/nginx-load-balancer-conf
