+++ ingress/controllers/nginx/examples/rc-ssl.yaml
@@ -1,8 +1,7 @@
-
 apiVersion: v1
 kind: ReplicationController
 metadata:
-  name: nginx-ingress-3rdpartycfg
+  name: nginx-ingress-controller
   labels:
     k8s-app: nginx-ingress-lb
 spec:
@@ -15,12 +14,8 @@ spec:
         k8s-app: nginx-ingress-lb
         name: nginx-ingress-lb
     spec:
-      volumes:
-      - name: dhparam-example
-        secret:
-          secretName: dhparam-example
       containers:
-      - image: gcr.io/google_containers/nginx-third-party:0.4
+      - image: gcr.io/google_containers/nginx-ingress-controller:0.4
         name: nginx-ingress-lb
         imagePullPolicy: Always
         livenessProbe:
@@ -51,10 +46,9 @@ spec:
           hostPort: 4444
         - containerPort: 8080
           hostPort: 9000
-        volumeMounts:
-        - mountPath: /etc/nginx-ssl/dhparam
-          name: dhparam-example
+        # to configure ssl_dhparam http://nginx.org/en/docs/http/ngx_http_ssl_module.html#ssl_dhparam
+        # use the dhparam.sh file to generate and mount a secret that containing the key dhparam.pem or
+        # create a configuration with the content of dhparam.pem in the field sslDHParam.
         args:
-        - /nginx-third-party-lb
-        - --tcp-services-configmap=default/tcp-configmap-example
+        - /nginx-ingress-controller
         - --default-backend-service=default/default-http-backend
