+++ ingress/controllers/nginx/lua/error_page.lua
